+++ src/FsPickler.Json/AssemblyInfo.fs
@@ -3,9 +3,9 @@ open System.Reflection
 
 [<assembly: AssemblyProductAttribute("FsPickler")>]
 [<assembly: AssemblyCopyrightAttribute("© Eirik Tsarpalis.")>]
-[<assembly: AssemblyVersionAttribute("2.1.0")>]
-[<assembly: AssemblyFileVersionAttribute("2.1.0")>]
+[<assembly: AssemblyVersionAttribute("2.3.0")>]
+[<assembly: AssemblyFileVersionAttribute("2.3.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "2.1.0"
+    let [<Literal>] Version = "2.3.0"