+++ tests/FsPickler.Tests/GenericTests.fs
@@ -242,7 +242,11 @@ module ``Generic Tests`` =
     let ``2. Clone: array byte`` () = checkArray<byte> ()
 
     [<Test; Category("Clone")>]
-    let ``2. Clone: array enum`` () = checkArray<Enum> () ; checkArray<CharEnum> ()
+    let ``2. Clone: array enum`` () = 
+        checkArray<Enum> () ; 
+        // provisional: until mono bug is fixed
+        // https://bugzilla.xamarin.com/show_bug.cgi?id=40568
+        if not runsOnMono then checkArray<CharEnum> ()
 
     [<Test; Category("Clone")>]
     let ``2. Clone: cached array`` () =